pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "daily-tracker"
version = "0.0.0"
description = "An application for keeping track of tasks throughout the day."
authors = ["Bilbottom"]
readme = "README.md"
repository = "https://github.com/Bilbottom/daily-tracker"
packages = [
    {include = "daily_tracker"},
    {include = "core", from = "daily_tracker"},
    {include = "integrations", from = "daily_tracker"},
]

[tool.poetry.scripts]
daily-tracker = "daily_tracker.main:main"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = ">=9.4.0"
python-dotenv = ">=0.21.1"
pyyaml = ">=6.0"
requests = ">=2.28.2"
slack-sdk = ">=3.20.0"
streamlit = "^1.30.0"
ttkthemes = "^3.2.2"
pywin32 = [
    {platform = "win32", version = "306"}
]
appscript = [
    {platform = "darwin", version = "1.2.2"}
]

[tool.poetry.group]
dev.optional = true
test.optional = true
ide.optional = true

[tool.poetry.group.dev.dependencies]
coverage-badge = "^1.1.0"
pre-commit = "^3.6.2"
pylint = "3.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"

# Packages just for IDE integration
[tool.poetry.group.ide.dependencies]
black = "*"


[tool.pytest.ini_options]
addopts = "--cov=daily_tracker --cov-fail-under=10"
testpaths = ["tests"]


[tool.black]
target_version = ["py311"]
line-length = 80


[tool.isort]
profile = "black"
line_length = 80
known_local_folder = "daily_tracker,core,integrations,utils"


[tool.pylint.format]
max-line-length = 80

[tool.pylint.MASTER]
ignore-paths = "^tests/.*$"
